Concluindo o crud

+++ src/pages/Index.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react'
 import { getProdutoApi } from '../service/Api';
 import { deleteProdutosApi } from "../service/Api";
-import Container from '../componentes/Container'
-import Cabecalho from '../componentes/Cabecalho'
-import Menu from '../componentes/Menu'
+// import Container from '../componentes/Container'
+// import Cabecalho from '../componentes/Cabecalho'
+// import Menu from '../componentes/Menu'
 import Body from '../componentes/Body'
 import Rodape from '../componentes/Rodape'
 import Produtos from '../componentes/Produtos'
@@ -11,6 +11,7 @@ import Form from '../componentes/Form';
 import MenuProduto from '../componentes/Menu/MenuProduto';
 import '../App.css'
 import '../assets/styleGlobal.css';
+import Layout from '../componentes/Layout';
 
 function Initial() {
   const [listas, setProdutos] = useState([])
@@ -22,6 +23,11 @@ function Initial() {
   // Deletar produtos
   const deleteProduto = async (id) => {
     console.log('id', id)
+    const confirmDelete = window.confirm("Tem certeza de que deseja deletar este produto?");
+  
+    if (!confirmDelete) {
+      return; // Cancela a exclusão se o usuário não confirmar
+    }
     try {
         await deleteProdutosApi(id);
         await getProduto();
@@ -38,6 +44,7 @@ function Initial() {
 
    // Ciclo de vida do componente
   useEffect(() => {
+    console.log('Atualizando')
     getProduto();
   }, []);
 
@@ -54,24 +61,17 @@ function Initial() {
     setUpdate(true);
   }
 
-  
-
- console.log("MUDANÇA NO USER_EFFECT",listas)
-
   return (
-    <>
-      <Cabecalho />
-      <Menu />
-      <Body>
-        <MenuProduto show={showList} act={cadastro} />
+    
+      <>
+        {/* <MenuProduto show={showList} act={cadastro} /> */}
         {showList ? (
           <Produtos produtos={listas} deleteProduto={deleteProduto} editForm={editForm} /> 
         ) : ( 
-          <Form form={form} setForm={setForm} update={update}/>
+          "Talvez Colocar essa lista de volta"// <Form form={form} setForm={setForm} update={update}/>
         )}
-      </Body>
-      <Rodape />
-    </>
+      </>
+    
   )
 }
